added some more tests, Card, Game; slightly refactored board
