+++ bin/gamesDir/risk/maps.js
@@ -1,113 +1,4 @@
-"use strict";
-
-class Region {
-   constructor(name, parent, neighbours) {
-      this._name = name
-      this._neighbours = neighbours
-      this._parent = parent
-      this._owner = null
-      this._units = 0
-   }
-
-   get name() {
-      return this._name
-   }
-
-   get neighbours() {
-      return this._neighbours
-   }
-
-   get owner() {
-      return this._owner
-   }
-
-   get units() {
-      return this._units
-   }
-
-   set owner(newOwner) {
-      this._owner = newOwner
-   }
-
-   set units(newUnits) {
-      this._units = newUnits
-   }
-}
-
-class Continent {
-   constructor (name, bonus) {
-      this._name = name
-      this._bonus = bonus
-      this._territories = {}
-   }
-
-   get bonus() {
-      return this._bonus
-   }
-
-   get name() {
-      return this._name
-   }
-
-   _addRegion(key, region) {
-      this._territories[key] = region
-   }
-
-   isOwned(player) {
-      for (let territory in this._territories) {
-         if (this._territories[territory].owner != player) {
-            return false
-         }
-      }
-      return true
-   }
-
-}
-
-class Board {
-   constructor() {
-      this._map = {}
-      this._continents = {}
-
-      for (let key in continents) {
-         this._continents[key] = new Continent(
-            key,
-            continents[key].bonus
-         )
-         for (let placeIndex in continents[key].territories) {
-            let place = continents[key].territories[placeIndex]
-            let neighbours = map[place].neighbours
-            let r = new Region(place, this._continents[key], neighbours)
-            this._map[place] = r
-            this._continents[key]._addRegion(place, r)
-         }
-      }
-   }
-
-   get places() {
-      return Object.keys(this._map)
-   }
-
-   getRegion(region) {
-      return this._map[region]
-   }
-
-   getContinentBonuses(player) {
-      let bonus = 0
-      for (let cont in this._continents) {
-         if (this._continents[cont].isOwned(player)) {
-            bonus += this._continents[cont].bonus
-         }
-      }
-      return bonus
-   }
-
-   getNeighbours(region) {
-      return this._map[region] ? this._map[region].neighbours : null
-   }
-}
-
-let map = {
+const Regions = {
    Afghanistan: {
       neighbours: [
          'Ukraine',
@@ -487,7 +378,7 @@ let map = {
    }
 }
 
-let continents = {
+const Continents = {
 
    'North America': {
       bonus: 5,
@@ -568,4 +459,5 @@ let continents = {
    }
 }
 
-module.exports = Board
+exports.Regions = Regions
+exports.Continents = Continents
